refactor : 이슈 레포지토리 Spring data JPA 적용
